requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
